requirements.txt
streamlit>=1.20
pdfplumber>=0.7.5
python-docx>=0.8.11
scikit-learn>=1.0
